Allow fixture / test reuse by making factories
This change was necessary in order to make initialize idempotent.

Although the previous implementation worked for some simple cases,
every instance of a test / fixture was the same node in the tree. This
probably would have lead to some weird, unexpected loops in graphs if a
fixture was reused.

Now, a new Fixture / Test is generated every time it is called inside
of a suite.